\name{opencl}
\alias{opencl}
\alias{opencl.platforms}
\alias{opencl.platform}
\alias{opencl.devices}
\alias{opencl.device}
\alias{print.opencl.platforms}
\alias{print.opencl.platform}
\alias{print.opencl.devices}
\alias{print.opencl.device}
\alias{property}
\alias{property.opencl.platform}
\alias{property.opencl.device}
\title{opencl}
\description{
    OpenCL objects.
}
\usage{
    property(x, name)
    opencl.platforms
    \method{print}{opencl.platforms}(x, \dots)
    opencl.platform
    \method{print}{opencl.platform}(x, \dots)
    \method{property}{opencl.platform}(x, name)
    opencl.devices
    \method{print}{opencl.devices}(x, \dots)
    opencl.device
    \method{print}{opencl.device}(x, \dots)
    \method{property}{opencl.device}(x, name)
}

\arguments{
\item{x}{input object }
\item{name}{property name}
\item{...}{other stuff}
}
\author{
Michael Budiansky
quadrivio@7402.org
}

\keyword{multiblas}
